Vsip report restructure (#835)
